agrosilo-ts-pipeline/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
motor==3.4.0
pymongo==4.6.3
pydantic==2.9.2
python-dotenv==1.0.1
pydantic>=2.8
numpy>=2.0
pandas>=2.2
reportlab>=4.2
python-dateutil>=2.9